Edit search routes.
Add function and route to retrieve number of likes given an activity.

Add function to retrieve anime results from my database.

Edit server response for movies and tv to send anime ID data along with the "This is an anime" message.

Edit search function to display search results for movies, tv, and anime.